using muParser (double type) and gLib2D as packed libraries, reorganized files

+++ function.h
@@ -8,7 +8,7 @@ class Function
 {
 private:
     mu::Parser parser;
-    float cx, cy;
+    double cx, cy;
     bool valid;
     
 public:
